Add confirm dialog

+++ 04-cart-app-routing/src/app/components/cart-app/cart-app.component.ts
@@ -4,8 +4,11 @@ import { Product } from '../../models/product';
 import { CatalogComponent } from "../catalog/catalog.component";
 import { CartItems } from '../../models/cartItems';
 import { NavbarComponent } from '../navbar/navbar.component';
-import { RouterOutlet } from '@angular/router';
+import { Router, RouterOutlet } from '@angular/router';
 import { SharingDataService } from '../../services/sharing-data.service';
+import Swal from 'sweetalert2'
+
+
 
 @Component({
   selector: 'cart-app',
@@ -20,15 +23,14 @@ export class CartAppComponent implements OnInit{
   total: number = 0;
   showCart:boolean = false;
 
-  constructor(private service:ProductService,private sharingDataService: SharingDataService){
+  constructor(
+    private router: Router,
+    private service:ProductService,private sharingDataService: SharingDataService){
 
 
   }
 
   ngOnInit(): void {
-
-    this.products = this.service.findAll();
-
     this.items = JSON.parse(sessionStorage.getItem('cart')!) || []
     this.calculateTotal();
     this.onDeleteCart()
@@ -56,6 +58,15 @@ export class CartAppComponent implements OnInit{
       }
       this.calculateTotal();
       this.saveSession();
+      this.router.navigate(
+        ['/cart'],
+        {state: {items:this.items,total:this.total}}
+      )
+      Swal.fire({
+        title: "Shooping Cart",
+        text: "Nuevo producto agregado al carro!",
+        icon: "success"
+      });
     } )
   }
 
@@ -63,14 +74,44 @@ export class CartAppComponent implements OnInit{
   onDeleteCart():void{
 
     this.sharingDataService.idProductEventEmitter.subscribe(id =>{
-      console.log('Se ha ejecutado el evento ' + id)
-      this.items = this.items.filter(item=> item.product.id !==id );
-      if(this.items.length === 0 ){
-        sessionStorage.removeItem('cart');
-        sessionStorage.clear();
-      }
-      this.calculateTotal();
-      this.saveSession();
+      Swal.fire({
+        title: "Esta seguro que desea eliminar?",
+        text: "Cuidado el item se eliminara del carro de compras!",
+        icon: "warning",
+        showCancelButton: true,
+        confirmButtonColor: "#3085d6",
+        cancelButtonColor: "#d33",
+        confirmButtonText: "Si,eliminar!"
+      }).then((result) => {
+        if (result.isConfirmed) {
+
+          this.items = this.items.filter(item=> item.product.id !==id );
+          if(this.items.length === 0 ){
+            sessionStorage.removeItem('cart');
+            sessionStorage.clear();
+          }
+          this.calculateTotal();
+          this.saveSession();
+          this.router.navigateByUrl('/',{skipLocationChange:true}).then(() =>{
+            this.router.navigate(
+              ['/cart'],
+              {state: {items:this.items,total:this.total}}
+            )
+          })
+
+
+
+          Swal.fire({
+            title: "Eliminado!",
+            text: "Se ha eliminado el carro de compraas.",
+            icon: "success"
+          });
+        }
+      });
+      
+
+
+    
 
     })
  
@@ -85,4 +126,5 @@ export class CartAppComponent implements OnInit{
   openCart(){
     this.showCart= !this.showCart
   }
+
 }
